requirements.txt
flask
flask-wtf
flask-sqlalchemy
flask-bcrypt
flask-login
flask-migrate
